removed useScreenSize - renamed Hook files with a suffix
