+++ client/src/shared/hooks/api/useQuizDataFetch.hook.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { URLS } from "../../../configs/Links";
-import { HttpService } from "../../services/httpService";
+import { HttpService } from "../../services/httpService.service";
 
 const httpService = new HttpService();
