--- client/src/shared/hooks/useScreenSize.tsx
@@ -1,32 +1,0 @@
-import { useEffect, useState } from "react";
-
-interface ScreenSize {
-  width: number;
-  height: number;
-}
-
-const useScreenSize = () => {
-  const [screenSize, setScreenSize] = useState<ScreenSize>({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
-  useEffect(() => {
-    const handleResize = () => {
-      setScreenSize({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    };
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  return screenSize;
-};
-
-export default useScreenSize;
